# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dangnhap": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Đăng nhập tài khoản",
                "operationId": "32ec4ae4caf2da34eef6cb998a613204",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Email",
                                    "Matkhau"
                                ],
                                "properties": {
                                    "Email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "truongminhthien222004@gmail.com"
                                    },
                                    "Matkhau": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "thien123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/dangki": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Đăng kí tài khoản",
                "description": "Thêm mới một người dùng",
                "operationId": "d470b0e0102d39da4757c6cac26d996f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Hovaten",
                                    "Email",
                                    "Diachi",
                                    "SDT",
                                    "Matkhau"
                                ],
                                "properties": {
                                    "Hovaten": {
                                        "type": "string",
                                        "example": "Trương Minh Thiện"
                                    },
                                    "Email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "truongminhthien222004@gmail.com"
                                    },
                                    "Matkhau": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "thien123"
                                    },
                                    "SDT": {
                                        "type": "string",
                                        "example": "0354895845"
                                    },
                                    "Diachi": {
                                        "type": "string",
                                        "example": "294 -296 Đồng Đen - Quận Tân Bình - Hồ Chí Minh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Người dùng đã đăng kí thành công tài khoản",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Cập nhật thông tin người dùng",
                "description": "Cập nhật thông tin người dùng",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Người dùng đã được cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User updated successfully"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người dùng"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Xóa người dùng",
                "description": "Xóa một người dùng",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Người dùng đã bị xóa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người dùng"
                    }
                }
            }
        }
    },
...
}